Update nms.model.train.py
Updated train function tracks learning improvement according to some tolerance.
Train function takes a tolerance parameter which is the tolerated decrease in loss over successive epochs that we consider sufficient to update the model. 
Function also takes a patience parameter which is the number of successive epochs with loss less than tolerance that will stop training.